remove old ch_frb_rpc.cpp; add test-l1-rpc.cpp for demo
